GOFR
Following Program uses gofr for inserting entries onto a mysql database and with the help of /customer we can see all the data inside the database. You can manually search for any particular persons data as i have created an endpoint named customer/search/{name} for it . For inserting data you can use the curl queries to input the data onto the database. Additionally provided a frontend page that displays all the user data with the help of GUI